scanner: Move ArtifactoryStorage's "read(Package, ScannerDetails)" to a separate class
Move it to a new abstract class so it can be shared with other (file-based)
storage implementations.
